oci: do not store index twice (#51387)
This is the least invasive fix to ensure OCI build caches do not fetch
an index twice and store it under `_v2` and `_v3` in cache, while also
not fetching it yet a third by dropping the `_v*` suffix.

Databases can get quite big, so this is pure overhead for OCI, since 
the v2 and v3 suffixes apply to URL based build caches only.

The v2 related code can simply be removed together with this small
patch in a future Spack release.