solver: account for language preferences on reused specs (#51417)
Due to performance and modeling reasons, the specs that are
reused in concretization are represented in the ASP problem
with their pure build dependencies trimmed.

On `develop` this means that any preference of the form:
```yaml
packages:
  c:
    prefer:
    - llvm
```
is "disregarded" for purely reused specs, since the solver
is not aware of the compiler nodes used for them.

This commit fixes the semantics of virtual preferences so
that they are respected also for reused specs. It does so
without representing their compilers as nodes - so the modeling
of the DAG is not steered towards different optimal solutions
by the missing compilers.

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>