solver: penalize language virtuals for each use (#51314)
This commit changes how penalties for language virtuals are computed
so that a penalty is taken each time a non-preferred compiler is used.

Using this environment as an example:
```yaml
spack:
  specs:
  - "fftw~mpi %c,fortran=llvm"
  packages:
    c:
      prefer: [gcc]
    cxx:
      prefer: [gcc]
    fortran:
      prefer: [gcc]
```
**Before the PR**
Every node uses `clang` as a compiler, since the penalty is taken per
provider, and we force `clang` on `fftw`.

**After the PR**
Only `fftw` uses `clang`, since a penalty is taken per provider use, 
so switching other nodes to `clang` is suboptimal (since `gcc` is preferred).

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>