bugfix: don't stop early when other spack process is installing (#51539)
The installer in #47590 missed a loop condition and wasn't properly waiting on tasks
requeued as `BuildStatus.INSTALLING`. This means another Spack process is working
on that task, and we should wait and not bail early.

There is also a condition in the installer where we discover for the first time that a
package is installed in cleanup. We need to update the installed dependencies lists of
dependents when that happens.

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>