new_installer: display help (#51557)
Show available commands in the last line of the new installer's terminal
UI.

Currently just `/`, `v`, `n/p`. I don't think a big help screen is
needed really. What's not displayed is Esc and q, but I think that's
natural. The reason to leave them out is to make the line compact, but
also to avoid having to communicate that q does not quite the installer,
but rather only returns from logs to overview.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>