stage.py: fix type erasure issues (#51603)
The StageComposite class is a black box for type checkers.

It's an object that takes multiple stage instances, and provides a
single interface for all, delegating function calls to each instance,
like `create()` and `fetch()`.

However, the way it's implemented is curious: it subtypes `list` and is
implemented in a very generic way using a Composite object.

This PR removes all that and makes the interface explicit, so that type
checkers can find issues.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>