asp.py: do not sort pkg.dependencies and versions (#51632)
The dictionary pkg.dependencies has a deterministic order based on the
package.py contents, and package_dependencies_rules is called on a
sorted list of packages.

Together that gives a deterministic condition id that only depends on
the order of directives in package.py. That should be enough for
concretization cache to work.

Removal of the sort calls reduces the runtime of the setup phase of the
solver by 6%.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>

* asp.py: drop sorting version satisfaction facts

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>

---------

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>